requirements.txt
asgiref==3.5.0
autopep8==1.6.0
black==22.1.0
certifi==2021.10.8
cffi==1.15.0
click==8.0.4
cryptography==36.0.2
Django==4.0.3
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-jwt==1.19.2
flake8==4.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
pytz==2021.3
sentry-sdk==1.5.8
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing-extensions==4.1.1
urllib3==1.26.9
